Split presentation logic from auth helper
